desktoppackage/configuration: use value roles for action comboboxes
The comboboxes for the containment actions configuration pages are
based on a role-based model, but only use these roles directly in the
comboboxes to set the text, manually iterating over the models to set
them based on value and reading from them to determine the value.

Instead, we can set the value role then use indexFromValue() to set them
and currentValue to read the plugin names. This simplifies the code, and
also eliminates calls to the containment model's get() function, which
was causing runtime error messages.