PanelConfiguration: move two "Add [thing]" buttons into a menu
This dialog is intended to be very narrow, yet its header has a label
and two buttons, one somewhat long. In English this works out okay, but
when translated into other languages, it's possible for the labels to be
very long aand cause the dialog to balloon in size, which looks really
bad.

Instead, collapse the two buttons into a menu, which ensures there's
enough space with all languages.