Folder View: Don't show wrong context menu when left-and-right-clicking
On devices with physical left and right buttons, it's possible to
right-click while holding down the left button. When doing this on the
desktop, you get the wrong context menu, because Folder View code was
inappropriately handling the event rather than falling back to the
desktop context menu plugin.

To fix this, explicitly handle that case and don't accept the event,
allowing it to fall back properly.

BUG: 501856
FIXED-IN: 6.4.0