desktoppackage/configuration: call applet's own save function first
Currently, saving the configuration first saves the cfg_ properties,
then calls the applet's own saveConfig() function. This is inconvenient
in case applets might want to manage some of the cfg_ properties by
themselves – they not only need to implement managing the data, but also
the actual saving. (One case where this might be useful is unordered
lists; the automatic detection of changed properties assumes that order
matters and we cannot easily change this for reasons of backward
compatibility). If saveConfig() was called first, the applet
configuration could simply set the cfg_ properties there, and have
AppletConfiguration handle the saving.

This change switches the order and updates an outdated comment.