applets/taskmanager: demote bounds check to a debug
This can be useful for debugging, but the condition is harmless for a
normal user. Warning is probably too severe a severity for this, as it
will pollute the journal log with tons of un-actionable lines when there
are any miebehaving apps that set out-of-bounds progress values. Demote
it to debug.