desktoppackage/configuration: remove support for loading kcms
In the past, kcms could be loaded as configuration pages in applet
configuration dialogs. This is no longer used (instead, we provide
buttons or menus to open systemsettings on the corresponding page) and
probably doesn't work reliably anymore, and is thus being removed from
libplasma. This makes it unnecessary to keep support for it around in
the applet configuration dialog.

This change removes the handling for kcms as applet configuration pages.