desktoppackage/configuration: remove support for includeMargins
The includeMargins property in Plasma 6 was only ever applied to
containment configuration pages, and in practice only set for
ContainmentConfigurationAppearance. Now that containment configuration
pages are also implemented as KCMUtils components, there is no need for
this anymore, and support for it is being removed from libplasma.

We can also simplify the frontend a bit by getting rid of the handling
for this property.