desktoppackage/configuration: only save wallpaper config if changed
b542ac85 changed the plasmoid configuration dialogs to not prompt for
saving if the original state was restored, but some pages that do not
use the standard methods of configuration saving still prompt for saving
if they were modified at all, no matter the present state. The wallpaper
configuration with its embedded plugin configuration is one such case.

This change implements a check for wallpaper configuration that compares
the current and stored configuration and only prompts for saving if the
new configuration is actually different.

Co-authored-by: Méven Car <meven@kde.org>