desktoppackage/ConfigurationShortcuts: use FormLayout
Most applet config pages use a FormLayout or are based around a ListView
or GridView. The Shortcuts configuration page is an exception, it uses
a simple ColumnLayout. As a consequence, the contents (a helper text
line and a KeySequenceItem) are not centered and end up with somewhat
awkward margins. This is atypical for such dialogs in Plasma, especially
now that even narrow-mode FormLayouts will typically have some form of
margin.

Instead, use a FormLayout here as well, and use the helper text contents
as a label for better visual consistency with other pages