desktoppackage/AppletConfiguration: clear state when discarded
When a user makes a change that causes a configurationChanged signal to
be sent, switches to a different config page, and clicks "Discard" on
the confirmation dialog, the internal state is not cleared, and from
then on modification tracking no longer works (until the user presses
Apply or OK on the config dialog) - as long as any setting was changed,
it will always treat it as modified, even if the original state is
restored. Very few of our own applets still send this signal, as most
that needed it have been ported to monitoring the unsavedChanges
property instead, but we cannot expect third-party widgets to do this.

Instead, clear the variable that tracks this state on discarding the
dialog.