Add new scroll mode to cycle through windows of a single task
New take on https://invent.kde.org/plasma/plasma-desktop/-/merge_requests/2434

Currently, there are two options for scrolling behavior: disable it entirely,
or scroll through all tasks. This commit adds a third option: only scroll
through the windows of the task under the cursor. When this mode is selected:

- If the task has multiple windows, scrolling will cycle only through those windows
- If the task has a single window, scrolling will simply raise it
- If the task is a launcher with no open windows, scrolling will have no effect

BUG: 461481
FIXED-IN: 6.4