applets/minimizeall: don't pretend to be Peek at Desktop in edit mode
The Minimize All and Show Desktop widgets cleverly share code and are
capable of undoing one another's actions when invoking the other's
functionality from the context menu.

However this becomes less useful while in Edit mode, which invokes "Peek
at Desktop" automatically. Then both widgets claim to be "Peek at
Desktop" and it can be hard to tell which one is which as you make
changes. So let's not do that.

BUG: 503624
FIXED-IN: 6.4.0