FolderModel: When adding files, keep the current sorting
The sort mode got changed to unsorted if any file was added.
This was probably due to the infinite while loop that would happen
later when we're looking for a new blank spot from an index
that does not exist: This would return an empty QVariant that is
false by default.

We need to check if the index is valid. If it's not, we know
it's a new index and can create one, so it is a blank spot.

BUG: 482470
FIXED-IN: 6.3.6


(cherry picked from commit 03ef2343212bbc4be274b1b37788186df03a8cb3)

10260acb FolderModel: When adding files, keep the current sorting
e64705c8 Fix comment, combine ifs

Co-authored-by: Akseli Lahtinen <akselmo@akselmo.dev>