containments/desktop: fix reopening Location config page
When the radio button for the Places option is selected, the combobox is
enabled and the path selected there is marked as the new path for the
widget. This is generally reasonable, but breaks when opening the
configuration page with anything except the first entry selected - the
combobox is enabled before the correct value is set, which causes it to
overwrite the correct value with the default combobox value, losing the
user's configuration.

Instead we just set the value before enabling the combobox and
everything works out fine.

BUG: 454889
FIXED-IN: 6.4.0