containments/desktop: make empty filter pattern consistent
When the applet is loaded, it defaults to an empty pattern ("") and
m_filterPatternMatchAll to true. But when a filter pattern is set in the
config dialog, then set to empty again, m_filterPatternMatchAll is false
(until Plasma is restarted and it falls back to its default state). This
creates an inconsistency where an empty pattern sometimes matches all
and sometimes it matches exactly nothing, which is confusing.

Matching only empty names does not seem to be useful ever, so let's
always consider it to match all. That way, users who want to show or
hide by file type don't have to worry about setting a glob in the file
name pattern field.