desktoppackage/AppletConfiguration: skip separators on GridViewKCM
The Applet configuration page only shows titlebar/buttonbar separator
lines is the page has the flickable property set. Unlike ScrollViewKCM,
GridViewKCM sets this property, so applet config pages based on it have
an extra line separating the titlebar from the header (or, if set, the
footer). This goes against the general dialog-like aesthetic of the
applet configuration pages.

Instead, don't draw the lines for GridViewKCMs. We also provide a
default false value for the visible property to silence an error with
Digital Clock's configTimeZones.