desktoppackage/AppletConfiguration: save FolderView containment
AppletConfiguration still has a conditional for containment config pages
from when they worked differently from other config pages. For the
standard containment pages (Appearance and Actions) this does not matter
as they don't use the cfg_ mechanism and both methods work out to the
same result. But for the FolderView config pages this breaks, as they
use cfg_ properties and don't handle the saving themselves.

Instead, just use the same path whether it's a containment or not - that
path works for everything now.