kcms/mouse: add ContextualHelpButton for on-button scrolling
Since Plasma 6.2, we allow users to toggle the libinput on-button
scrolling option from the mouse kcm. This function is intended for
scroll-wheel emulation on devices that do not have one, such as pointing
sticks, although some users may enjoy it on other devices as well. Users
often come looking for something more akin to Windows' auto-scrolling
feature, which we do not support. This leads to confusion, as its
behavior is substantially different from the feature they are looking
for, and there are significant downsides for some applications to having
this enabled (as it disables all middle-click dragging functionality).

Let's add a helper button that makes clear (a) what the downside of
enabling this option is and (b) what it is primarily intended for, so
users don't think it's a botched implementation of auto-scroll.

CCBUG: 504133


Co-authored-by: Nate Graham <nate@kde.org>