kcms/mouse,touchpad: get rid of tooltips
The mouse and touchpad kcms have tooltips for almost all entries, unlike
our other kcms. Almost all of them are just rephrasings of the visible
test and are therefore unnecessary.

This change gets rid of them. Two of the tap-and-drag related options
need some adjustments as they are not necessarily clear without the
tooltip. For tap-and-drag, we add a contextual help button with an
expanded version of the previous tooltip, and "tap-and-drag" lock is
rephrased to be more explicit about what it does.