applets/kickoff: make menu button labels always reflect what's in there
Right now the menu button label says "Power" when it's showing power
actions, but it shows "Leave" when it's showing session actions, and it
shows "More" when it's showing both power and session actions, and
nothing else on the toolbar.

Instead of getting fancy here and being inaccurate, just reflect what's
in the menu.

BUG: 499750
FIXED-IN: 6.4.0