applets/kicker: fix and simplify runner keyboard navigation
When the runner columns are shown, kicker's keyboard navigation breaks
in two ways: immediately moving to the right from the search field leads
to a selection in two columns, as the currentIndex is not reset, and the
whole thing becomes somewhat broken if the first column is not visible
(i.e. the search only matches in later runners). This is because
keyboard  navigation iterates through all possible columns and attempts
to find the right one, but this is complex and it often doesn't do a
good job of not taking invisible things into account.

Instead, we can get rid of the manual looping and tracking - qml already
computes a list of all the visible columns anyway for the
visibleChildren property. We just need to exclude the Repeater and then
we can directly access the correct column. We also clear the
currentIndex as necessary to fix the first issue.