containments/desktop: redetermine item on click
In case of touch events, Folder View already resets the hovered item to
the actual click position, as touch does not cause hover events. There
is another case where a click on a non-hovered element can happen,
namely if the containsMouse determination is blocked by an open context
menu. If the user does not move the pointer after opening a context
menu, moving the pointer to another item, and closing the context menu,
left-clicking will leave the icon completely unaffected and right-
clicking will re-open the previous context menu, rather than open the
one for the clicked entry.

Instead, always take the delegate that was clicked to activate or open
the context menu, whether on touch or not. This matches the behavior in
Dolphin, where in the same situation the hover feedback is not updated,
but clicking still works normally.

BUG: 504765