applets/kicker: elide long primary menu names
Kicker has some logic to set minimum and maximum widths for the list
delegates, and elides the string if it's longer than the minimum length.
This works well for submenus, as Plasma dialog will resize when needed.
But the fullRepresentation popup will not resize automatically, it's
restricted to the size it has (and also cannot be resized by the user).
If a menu name does not fit, the delegate is clipped at the
representation's edge, but the delegate continues, and any popup from
there will start after the delegate's real (invisible) end leading to
bizarre gaps.

Instead, set the maximum size to the minimum size for the main menu
representation. The submenus still resize as normal.

BUG: 503143