applets/kicker: fix Plasma.dialog being inappropriately resized
Kicker reuses the current level's cascading popup when a new one is
opened, rather than creatinga completely new one. This usually works out
but fails if the submenu itself has a submenu - the submenu gets resized
correctly, but afterwards is resized again to its previous size. The
specific cause seems to be the subsubmenu hiding itself, if we instead
wait until it is destroyed on the next event loop clear, the resizing
doesn't happen.

So let's not hide it. This seems to be fast enough even on old computers
to not lead to  perceptible lingering subsubmenus, and even if it did
happen in some rare situations, this would seem preferable to having a
menu that's the wrong size long-term.

BUG: 503582


(cherry picked from commit bf2cfe5f31c2837966178fb951c3f5cb1d824501)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>