applets/kicker: fix and simplify representation size
The horizontal sizing of the full representation is rather complex and
due to circular dependencies somewhat buggy; in particular, searching
for something that yields multiple results columns then clearing the
search field will leave the menu column with the combined width of all
the columns, rather than shrinking to its normal size.

We simplify this by having the optimal widths bubble up from the
individual parts to the whole, rather than being circular in both
directions. This also allows us to re-enable dynamic sizing for the
(non-runner) full representation without the previous bugginess. (We
keep the maximum width somewhat more limited at 1.5 times the normal
width, as the menu is harder to operate when it is very wide, compared
to the submenus or runner columns). Along the way, we use optional
chaining to get rid of a runtime error that occasionally happens when
closing Kicker.

BUG: 477005
FIXED-IN: 6.4.0