applets/kickoff: fix error when accessing context action
Kicker actions return whether the launcher should be closed after the
action or not. Kickoff fails at its attempt to use the the information
though, as the plasmoiditem is not accessible by id from the singleton
nested menu component. It's already passed in as a property, but just
not used here. The attempt to access by id fails with an error. This
has no major negative consequences beside the error, but does make the
interaction feel less responsive as kickoff will stick around until the
app launches a window, which may take a while.

Instead, use the property as intended. We also add a check to avoid
automatically closing it when pinned; this seems like an oversight in
the original logic.

BUG: 496324
FIXED-IN: 6.4.0