applets/kicker: don't trigger hover on keyboard open
Kicker's MenuRepresentation uses MouseEventListener's positionChanged
to do its hover activation handling, and unlike standard QQC components'
hovered property, it activates immediately on spawn even if the mouse
was not actively moved. This makes keyboard use rather inconvenient, as
the behavior differs based on the position where the pointer is when the
user triggers the launcher.

Instead, we ignore the first positionChanged event on each opening of
kicker. This does not notably affect mouse use (the user has to move one
additional pixel when moving from icon to menu), but makes shortcut-
triggered keyboard use much more pleasant. In the medium term, it
probably makes sense to port this to standard components (ItemDelegate,
standard TriangleMouseFilter instead of an incomplete custom
implementation etc,), but given the complexity and lots of custom
behavior of kicker, this is likely a larger undertaking.

BUG: 423688
FIXED-IN: 6.4.0


(cherry picked from commit 09fbef3e6f651836f08484f1e9d6dbd85bf750cc)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>