applets/kicker: make runner columns horizontally scrollable
When additional runners are enabled and there are hits from many of
them, kicker's runner columns can easily overflow. This makes some of
them inaccessible and looks as broken as it is (including potentially
escaping the representation width).

Instead we limit the maximum width by screen width, and make the whole
thing horizontally scrollable if the content is larger than that.
Keyboard navigation is adjusted to make the focused element visible,
mouse use does not scroll by itself (as that is very disorienting).

BUG: 389146