applets/kicker: give runner listview focus
When pressing right from the search field, we focus the runner columns
which passes focus through several internal focus scopes ending at the
ScrollView that contains the ListView. We handle arrow navigation there
for some reason, but not activation, so trying to activate that item
will not work and fall back to the search field, which activates the
first column's first item, rather than the currently selected one. If
the user moves up or down this is fixed, as then focus actually lands on
the selected item.

Instead, we mark the listview as having focus inside its focus scope.
This passes focus correctly downward, so the item can be activated
immediately.