FolderItemDelegate: use label implicitHeight in frame size calculation
Assuming that a GridUnit was roughly the height of a line of text was
safe before d8510e54f0a79334a11b7215cc3b28fdbe15149b in Kirigami, but
afterwards it no longer is. Now the frame height is wrong when the font
has a substantially different size from that of 10pt Noto Sans.

Fortunately we don't need to do anything super complicated to fix this;
just use the implicitHeight of the label, which ensures that the height
is correct for everything.

BUG: 490402
FIXED-IN: 6.3.6