applets/kicker: rerun search on model refresh
When the underlying model changes, such as from an app being installed
or uninstalled, any active queries may be invalidated and have to be re-
run. Currently, this will just clear the query and leave the widget in
the default state. This is surprising and inconvenient, as the user has
to repeat their input to get back to what they were doing.

Instead, note the current search term on model refresh and re-enter it
for the user, restarting the query. As query changes are on a timer,
we require that clear is exposed in the backend, which is done in a
separate p-w commit.

BUG: 418088
FIXED-IN: 6.4.0


(cherry picked from commit 27bcbf6380d754646cbd14aed54459d872cb629f)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>