containments/desktop: fix minimum size for fullRepresentation in panel
The fullRepresentation FolderView in a panel can get stuck if the panel
is overly full where it swaps around its representations in a loop,
freezing Plasma. This is similar to a related bug affecting the notes
widget in bug 505132.

We can use the same strategy here: in case of being a fullRepresentation
in a panel, we set the Layout minimum sizes to -1 and it will force the
size in a way that makes it not drop back to the compactRepresentation.