desktoppackage: fix Image wallpaper configuration
The Image wallpaper plugin configuration uses the loaded WallpaperItem
object from the actual containment to do its preview manipulation. This
is accessed from the Image plugin through a magic context property
called wallpaper that is set somewhere (probably containmentconfigview
in p-w? who knows). The problem is that this context property is not
updated when the WallpaperItem is changed (through reloading a different
plugin), so the preview function only works when Image is the starting
wallpaper plugin, not if you change to it. And now that we make sure to
not write the PreviewImage property to disk, even applying the changes
fails if you switch to it.

To fix this, we set the correct object from the ContainmentItem when
loading the configuration, rather than rely on Image guessing the
relevant magic context property by itself. And as the inital properties
are set when the plugin is selected in the configuration dialog, but the
WallpaperItem created when we save it, we need to make it a binding or
it'll be outdated and null after we press apply. This makes both saving
work again, and clicking to Preview images work after you press apply.