Desktop: reload plugin combobox after model change
Requires:
- https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/5458
- https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/5596

Installing, uninstalling, or upgrading a wallpaper plugin causes a
model reset. However the plugin combobox's value is set imperatively,
and not updated after a model reset, so it becomes blank.

To alleviate this, connect to a new signal added to the model after a
reset happens, and refresh the combobox's contents and current index.

BUG: 501586
FIXED-IN: 6.4.1