containments/desktop: fix folder view clicking
53d67b9a0b82fff40b2057f1ceb74b529b943df5 made clicking use the same
logic to determine the selected item for mouse and touch, rather than
relying on hover events. It seems the logic for touch was broken in some
cirumstances, namely scrolled views and certain alignments, where the
coordinate systems do not align.

A similar issue was already fixed elsewhere in folder view with
509b655ef2edef04a13d9afaab6349911beaaeb1, so we reuse the approach here.

BUG: 497498
BUG: 505718
BUG: 504569
FIXED-IN: 6.4.1

----

Would be good if someone could test if this works with touch, as I don't have a functional touch device handy. Might be that this was also broken there (couldn't find a bug report there though) or doesn't apply.