applets/taskmanager: disable window highlighting by default
Part of https://invent.kde.org/plasma/plasma-desktop/-/issues/149

This effect can trigger full-screen-flickering when invoked on maximized
windows, which is distracting and disorienting. When there are multiple
such windows and you make it happen by moving the pointer over their
tasks, it can actually trigger a seizure in people sensitive to that
kind of thing, since the flickering happens at faster than 3 Hz.

The effect is therefore too dangerous to have on by default; disable it.

CCBUG: 469983

@teams/accessibility