Folder View: don't show label shadows on desktop popup either
Part of https://invent.kde.org/plasma/plasma-desktop/-/issues/149.

Currently we disable the shadows when the widget is not a containment,
which works for standalone widgets and popups of standalone widgets. But
it does not work for popups of the root containment version.

Check for that too by making the delegate aware of its surrounding
context.