🍒applets/kicker: activate favorites on enter/return
For menu entries, Space to activate does not work (it sends a space to
the search field), only Enter/Return, but for the favorites buttons on
the side, only Space works, Enter/Return do nothing.

Instead, allow Enter/Return to activate the item as well, so users can
use the same keys in all modes. We also switch the implementation to
immediately close kicker on activation using the keyboard, like is
already done when  clicking an item with the mouse; this makes operation
feel more responsive.

BUG: 505542
FIXED-IN: 6.4.2


(cherry picked from commit 3e3837856c18998503283c0d44d8bb60c25dd809)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>