kcms/tablet: Refresh the "Map to screen" option during reset
This is because the combo box's current index is decided by a function
call, and isn't a property binding. We have existing infrastructure for
this scenario, it just needs to be repurposed for the Display tab now.

BUG: 505768
FIXED-IN: 6.4.1


(cherry picked from commit c5fe5a3026e9c77a0ad301723754f07cbd5a575d)

Co-authored-by: Joshua Goins <josh@redstrate.com>