lockscreen: prevent immediate prompting for the password
As of Qt 6.8.2, the fix for QTBUG-127122 causes a positionChanged event
to be emitted when handling QEvent::HoverEnter.

The lockscreen creates a full screen MouseArea on every connected
screen, one of which will cover the current position of the mouse
pointer. This causes QEvent::HoverEnter and then the positionChanged
event to be raised immediately. The mouse position did not change.
Ignore the spurious (first) positionChanged event, show the password UI
as usual in response to further events.

BUG: 499637