fix(kcm_activities): replace continue_with with QFuture::then
This fixes use-after-free errors caused by an ownership issue with the
future watcher in `continue_with`, which was detectable in Valgrind and
also caused a segfault when the KCM was run with hardened_malloc; these
errors no longer occur on my system with this fix.

Also removed the obsolete `optional_view`, which can be replaced with
`std::optional`.

Signed-off-by: Daniel Hast <hast.daniel@protonmail.com>