FolderModel: If screen is not used, clear dropTargetPositions
If screen is not in use, we are not certain what our
current coordinates for the icons are, especially for desktop icons.

In these cases we should just clear the drop target positions and
return,
since this action only makes sense when the screen is in use
anyway.

This was also never cleared after dragging, except for the 10 second
timer.

This would cause edgecases where an icon is dragged around, something
causes rowsInserted to be emitted and we would reuse the drag
operation, which would move the icon around further.

Instead we should always clear it after any drag operation.

BUG: 490891

(cherry picked from commit 7fc477efadc783ba352bd3fbf11ee129fd237cfe)