FolderModel: Take layoutDirection into account with ordering
On desktop, user can align items to the left or right side of the view.
However when this is done, it would also reverse the orders of the items
which is incorrect:
If user has LeftToRight reading alignment, but aligns items to the
Right, the items would also use RightToLeft reading alignment.

This patch makes sure that we flip the reading alignment to correct one,
despite what alignment the view has.

BUG: 449069