kcms/tablet: Fix KCM settings disappearing when disconnecting a device
I added another check in onCountChanged to change our index in case a
device is removed. If we don't do this, it would show an empty settings
page that looks incredibly buggy (because it was buggy!)

An easy way to recreate this scenario without having multiple tablets is
to use something like my Tablet Simulator, open a fake input and then
close it.