kcms/tablet: Add button to configure built-in display
Configuring the built-in display for graphics tablet is quite common, so
I added a button to go directly to it's display settings. The heuristic
isn't perfect, but it's good enough until the ecosystem develops
something better.

If the tablet does not have a built-in screen, this button is hidden.

This also lays the groundwork for a better visualization of mappings to
the built-in tablet display, where the one we have now doesn't really
make sense.