kcms/tablet: Make pen tab scrollable
Unlike the other tabs, the base of the Pen one is an Item (not a Layout)
and I forgot to set it's implicit height. In certain configurations it
can extend past the minimum size of the window, making certain settings
inaccessible.

Now the implicit height is set, with the correct amount of padding too.