applets/kicker: correctly position search field
The search field in Kicker can be positioned wrong in a few cases,
pushing its very edge just outside the popup which looks wrong. This is
due to extra inappropriate spacing, resulting from two problems. The
first is that for top- and right-aligned panel applets, the state values
are incorrectly set and add superfluous spacing that is not present in
the left- or bottom-aligned base versions. The second issue is that in
case of an empty favorites sidebar, the search field (which is
positioned with anchors) adds a smallSpacing that it shouldn't, as the
content Row's spacing does not apply between two items if one of them is
not visible.

Instead, we remove both causes of extra spacing, making the search field
fit properly.

BUG: 507397
FIXED-IN: 6.4.4


(cherry picked from commit 2e1a5b38bc56ce594df08dbdf3d50b4c7b8bdf92)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>