applets/window-list: add main.xml to resources
The CMakeLists file for window-list does not include the KConfigXT
definition file. As a consequence, the file is not built into the applet
plugin, and the single configuration option does not work and causes
a runtime warning when the plugin is loaded.

Instead, include main.xml in the plugin definition.