FolderView: Simplify dir.onMove, make sure we dont create items outside of perStripe values
If the current column is same as perStripe, we are outside of the
creation area.

In these cases, lower the column value by one column, and the item
will be created in right spot.

Then we can move the item as expected.

Otherwise, creating new item after clicking right at the right edge of
the gridview would create it on the next row, even if there is space
left.

This patch also cleans up the code a bit and adds comments.

(cherry picked from commit 4c1529597dc80f1530dbb3c7c1241efc45d8df70)