applets/kicker: show runner results with no favorites
There is a complex dependency between the visibility of the runner
results and the favorites sidebar in kicker: if the sidebar is set to
invisible (due to having no contents), this is somehow propagated to the
runner search results, which also become invisible. I could not find
what exactly causes this dependency, but strongly suspect that the
complex anchoring layout and sizing is somehow involved, as a cleanup
that involves switching to less manual sizing and positioning resolves
the problem (!3148) as a positive side effect. But that is a more
invasive change that is not suitable for backporting and fixing the
issue in 6.4.

As a temporary workaround for the 6.4 branch only, we keep the sidebar
visible if empty while runner results are being displayed. It's still
at 0 width so only a very thin line is actually visible, and this
minor visual glitch seems preferable to having broken search
functionality when the user has no favorites. A separate MR (!3149) is
necessary or another glitched line will be displayed.

CCBUG: 507398