applets/kicker: remove unnecessary LayoutProxyItem
Kicker's sidebar uses LayoutProxyItems to rearrange the order of app and
system favorites depending on the panel edge. The separator line between
them also uses one, although it has a fixed position. This was
originally done for consistency, but it causes a bug: as the proxy
overrides the original's visibility, the separator is also visible when
there is only one favorite category, even though there is explicit code
intended to hide it in that circumstance. Furthermore, the benefit of
gaining a minor amount of consistency at the cost of having more and
uneccessary code is far from clear.

Instead, we just place the actual separator at the position it should
be, fixing the issue and reducing the code size (and saving a LPI's
worth of resources).


(cherry picked from commit acfd3105a2009299e169ef83f7a9809af286527b)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>