kcms/keys: make shortcut configuration keyboard-accessible
The list view that displays the shortcut-able actions for each category
does not itself take keyboard focus, and neither do its (unexpanded)
delegates. This means that unless the category only contains a single
action (which will then be expanded and show the KeySequenceItem and
possibly default shortcut checkbox), the keyboard shortcuts cannot be
configured with the keyboard alone, as there is no way to select and
unexpand an action delegate.

Instead, we allow the list view to take keyboard focus. To make the
interaction clear, we set the first entry as the currentItem when the
ListView receives focus, indicating that the user can now use arrow keys
to select the desired action.