applets/kicker: switch ItemListDelegate to required properties
Currently, ItemListDelegate has a mixture of accessing the model data
through the model property, through implicit lookups spread throughout
the code, and through a number of properties that are essentially just
accessors through the model (similar to required properties, but not
actually using them). This is somewhat confusing and makes it unclear
where exactly the data comes from, leading to errors such as the sorted
property not actually working (it's not a property of the model that's
passed to the delegate, but of the actual model item, so model.sorted is
always undefined).

This change switches all of these to explicit required properties and
qualifies the access where necessary. This makes the structure clearer
and qmllint happier (and we also access the correct sorted property
now).