kcms/tablet: Add a warning before configuring virtual tablet devices
There's quite a few ways to get graphics tablet drivers. One of the most
common is installing a user-space one, which usually comes with it's own
bespoke configuration tool that will interfere with our KCM here.

I added a warning to the top of the KCM for said devices, which can be
detected quite reliably. It's main purpose is to deter people from the
KCM unless they really know what they're doing. As I said before,
configuring the device in two places can really mess stuff up in
non-obvious ways.