applets/kicker: port ItemListDelegate to standard mouse handling
ItemListDelegate has rather custom mouse handling in a rather complex
MouseArea with cumbersome internal state tracking. Some of this is
relatively difficult to port to more standard handling, but we can
simplify it a fair bit by moving click handling and drag detection to a
TapHandler and DragHandler respectively.

This change implements that.