applets/kicker: separate hover blocking more cleanly
Kicker only activates hover events once the mouse was moved, so that
activating though the keyboard is predictable and does not focus
different items if the keyboard happens to be over them. The
implementation is not very clean however as the logic is spread out
between the ListView and the delegates. This creates unneeded
dependencies and makes it harder to port to less custom mouse handling
in the delegates.

Instead, we create a mouse area that just checks if the mouse is moved,
and then turns itself off. This allows us to do the logic there, and
just bind hoverEnabled of the children to turn it off and on. In the
short term this is a bit more code and an extra mouse area, but is
cleaner, more semantic, and will make it easier to remove the custom
MouseArea in the delegates later.