kcms/keyboard: Use ShortcutHelper::isDefaults in its intended place
Before this, `KCMKeyboard::isDefaults()` was calling
`m_shortcutHelper->isSaveNeeded()` among a number of designated
`isDefaults()` calls on other objects. This sits right below
`KCMKeyboard::isSaveNeeded()` where `isSaveNeeded()` functions
should actually be used.

Given that `ShortcutHelper::isDefaults` indeed exists, it's fair
to assume that it was meant to be used as part of the overall
`isDefaults()` function instead of `isSaveNeeded()`.
So fix the copy-and-paste mistake and use the function as intended.