applets/kicker: fix separators on sorted menu
Amends d5648c19293b617d6c6fa72ce724265bb1eae2cb

Kicker's delegate had some logic to hide separators when the menu is
sorted, as they would all pointlessly appear at the top. This consisted
of two parts, not loading them in the first place and setting their
height to zero. The first part didn't actually work as the wrong
property was accessed, so the separator was always loaded, and only
actually hidden by having zero height. The above commit fixed that. But
now that hiding the separators works properly, the separator in the main
menu between the Recent Applications/Files entries and the application
categories are also hidden. These do make sense even if sorted, and were
visible (albeit with wrong spacing) because a zero-height separator is
still visible if it's not at the edges of the list.

Instead, we make whether separators should be shown an explicit property
of the delegate (replacing sorted, which was not used for anyrhing else)
and set it explicitly for the main menu. This restores the previous
behavior (with correct spacing) with more explicit code.