applets/kicker: show placeholder on empty search result
Kickoff shows a simple placeholder message when a completed search
returns no matches. Kicker, however, just shows empty space. This is a
bit unclear, and as there is no indicator whether the search is still
running, it's also unclear whether there are actually no matches or the
file runner just needs a little more time.

Instead, show a placeholder very similar to the one for kickoff. It'll
only be shown once the search is actually complete.